feat: New memory model
That behaves more like a real EPOC device, in that it is permissive of
buffer overflows and lays things out in memory roughly like EPOC.

In terms of API changes, the AddrSlice class has been replaced by the
simpler Addr class, which does not track expected valid length.

The backing store for all variables is now the Chunk, which is how an
Addr can now be used to write into the storage of other Variables.

The ArrayValue class is removed as unnecessary (since there is no
concept of an array value type in OPL and the Chunk-based memory model
does not need it) and the array access semantics have been rolled into
Variable. The syntax for accessing an array item is now `array[idx]` rather
than `array()[idx]` and the taking the address of an array is now done with
`array:addressOf()` rather than `array()[1]:addressOf()`.

The Variable class is now decoupled from the underlying storage of the value
(which is done by Chunk) so Variables can now be constructed on-the-fly if
needed, providing a 'view' on the underlying storage. New API
`Runtime:addrAsVariable(addr, type)` facilitates this and replaces the old
`AddrSlice:dereference()` and `Runtime:addrFromInt()` APIs.

`Variable:setOnAssignCallback()` has been replaced by a dedicated `completion`
parameter in the iohandler `asyncRequest`.