Fix: Much improved date time handling
In particular stop using os.time() for anything, and instead introduce
utctime() API to iohandler. Epoc expects date/time calculations to be
treated as if times were UTC, ie that your clock is set "wrong" if you
are in another timezone, or to put it another way that your RTC is
always set to localtime rather than UTC + offset.

Also implement a variety of other date time functions.