fix: Crash when adding large directories (#239)
It looks like iOS 16 is far more agressive about limiting open file handles meaning we were running out of resources when monitoring large directory structures. This dials the number of open monitors down to 20 from 1000 and relies on manual refresh as a recovery mechanism. Hopefully we can find a better solution in the future.