fix: better menu scrollbar support
In particular in MENU calls with menus that have dividing lines in as
well as scrollbars, which necessitates supporting smooth scrolling as
opposed to only scrolling in multiples of the line height.

Also fixes several graphical glitches in menus with scrollbars, such as
submenu arrows not drawing in the correct location, performance, and
flicker.