refactor: Move common Swift and Lua into OpoLuaCore
This change introduces a new OpoLuaCore Swift package which contains the common Lua code and relevant Swift wrappers which were previously in 'swift'.

It's not at all clear to me if this is the correct approach, but it's here for discussion---this would make it easier to slowly pull common (non-AppKit/UIKit) code into a core package and allow us to incrementally move towards code that could be shared between native iOS and macOS apps. What I don't like about this change is that it moves the potentially very cross-platform Lua code firmly into a Swift package, potentially making it more fiddly to add other !Apple platforms in the future. Unfortunately, excepting symlinks, I can't think of a way to include the Lua 'src' directory in the OpoLuaCore package.