fix: Use common SIS file type identifier (#479)
This changes the SIS file type identifier from `org.opolua.sis` to a
more generic (but invented) `com.psion.sis` in an attempt to establish
an identifier that can be shared between different apps. In the near
term, the goal is to come up with an identifier that can be shared with
Reconnect.