fix: SCREEN was imposing incorrect maximum sizes
It was based on the default window's _current_ size, which is not
necessarily still the same as the screen size (which is the true
maximum). Even though PRINTs are rendered to the default window, so the
two things do really need to align, it's not mandated what order a
SCREEN and associated gSETWIN need to happen in so SCREEN cannot enforce
limits based on the current size of the window.