ci(qt): Use pre-built Qt for macOS builds (#589)
This change switches to using `jurplel/install-qt-action@v4` to install
Qt, rather than building it directly. It updates the workflow to use
named scripts for different Qt platform builds in anticipation that
these will all differ significantly.

It does not attempt to use embedded frameworks, but it likely should in
the future.