refactor: Adopt OpoLuaCore in the iOS app (#606)
This change updates the iOS app to use the OpoLuaCore package, ensuring
we’ll always be dog-fooding this package and ensures that we’re building
our core code for iOS _and_ macOS should we ever decide to push more
functionality into this common core and do a native macOS app
(unlikely). The code change is exclusively importing OpoLuaCore and
making properties, functions, and constructors public.